requirements.txt
elasticsearch
requests
requests_ftp
google-cloud
google-cloud-storage